fix & update
This update fixes several issues with tests and with the initial unreliable delta state update changes.

When explicitly setting state, it now is additive regarding flag states (i.e. it does not send outside of the tick event generated window).

Condensed the packet loss and standard network transform commonly shared code into a single NetworkTransformBase class.

Since state is not immediately pushed, had to update NestedNetworkTransforms test.

Updated several test projects that still had references to the UNet component which no longer exists in 2021.
