+++ com.unity.netcode.gameobjects/Tests/Runtime/NetworkTransform/NetworkTransformBase.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 07c2620262e24eb5a426b521c09b3091
+guid: ccdf46d0c73f8ac47921ec1be2772fac
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
